+++ quill/include/quill/Backend.h
@@ -21,7 +21,7 @@ QUILL_BEGIN_NAMESPACE
 
 /** Version Info **/
 constexpr uint32_t VersionMajor{6};
-constexpr uint32_t VersionMinor{0};
+constexpr uint32_t VersionMinor{1};
 constexpr uint32_t VersionPatch{0};
 constexpr uint32_t Version{VersionMajor * 10000 + VersionMinor * 100 + VersionPatch};